Backport https://github.com/opensearch-project/OpenSearch/pull/3075 to 2.x branch and resolve conflict in RecoveryState class
